feat: add NavBar component
