+++ src/app/layout.tsx
@@ -1,13 +1,6 @@
 import '@/styles/globals.css';
 
-import Link from 'next/link';
-
-import {
-  controlButton,
-  headControllers,
-  navBar,
-  titleBanner,
-} from './layout.css';
+import { NavBar } from './layout/NavBar';
 import { themeClass } from './style.css';
 
 export default function RootLayout({
@@ -19,17 +12,7 @@ export default function RootLayout({
     <html lang="ko">
       <body className={themeClass}>
         <main>
-          <nav className={navBar}>
-            <h1 className={titleBanner}>
-              <Link href="/">Conduit</Link>
-            </h1>
-            <ul className={headControllers}>
-              <li className={controlButton}>Home</li>
-              <li className={controlButton}>New Article</li>
-              <li className={controlButton}>Setting</li>
-              <li className={controlButton}>Profile</li>
-            </ul>
-          </nav>
+          <NavBar />
           {children}
         </main>
       </body>
