+++ src/app/layout/NavBar/index.ts
@@ -1,0 +1,1 @@
+export * from './NavBar';
